test(action-bar-module): add unit tests
